requirements.txt
aiofiles==0.7.0
anyio==3.2.0
astroid==2.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
h11==0.12.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
idna==3.2
isort==5.9.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
petname==2.6
pylint==2.8.3
PyYAML==5.4.1
rfc3986==1.5.0
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sanic==21.3.4
sanic-routing==0.6.2
sniffio==1.2.0
soupsieve==2.2.1
toml==0.10.2
ujson==4.0.2
uvloop==0.15.2
websockets==8.1
wrapt==1.12.1
